benchmark split. ALgorithm started
