+++ Makefile
@@ -4,10 +4,16 @@ GCCCXXFLAGS=-std=c++17
 CLANGHOME=./clang5
 CLANGCXX=$(CLANGHOME)/bin/clang++
 CLANGCXXFLAGS=-std=c++17 -stdlib=libc++
-LDFLAGS=-Xpreprocessor -fopenmp -lomp
 
-HEADERS=src/benchmark.h src/bin.h src/padded_vector.h src/interpolate.h src/util.h src/div.h src/lin.h
-SOURCES=src/search.cc
+OS := $(shell uname)
+ifeq ($(OS),Darwin)
+	LDFLAGS=-Xpreprocessor -fopenmp -lomp
+else
+	LDFLAGS=-fopenmp
+endif
+
+HEADERS=src/benchmark.h src/benchmark_datasets.h src/benchmark_utils.h src/binary_search.h src/padded_vector.h src/interpolate.h src/util.h src/div.h src/linear_search.h src/sip.h src/tip.h
+SOURCES=src/benchmark.cc
 
 .PHONY: run gdb clean perf
 
@@ -26,31 +32,30 @@ perf :
 clean:
 	rm -f ./searchbench ./debug ./dump
 
-## CLANG ######################
-####### Build Targets #########
-
-searchbench: CLANGCXXFLAGS += -O3 -DNDEBUG
-searchbench: $(SOURCES) $(HEADERS)
-	$(CLANGCXX) $(CLANGCXXFLAGS) $(SOURCES) -osearchbench $(LDFLAGS)
-
-debug: CLANGCXXFLAGS += -O0
-debug: $(SOURCES) $(HEADERS)
-	$(CLANGCXX) $(CLANGCXXFLAGS) $(SOURCES) -o$@ $(LDFLAGS)
-
-dump: src/dump.cc src/benchmark.h
-	$(CXX) $(CLANGCXXFLAGS) src/dump.cc -o $@ $(LDFLAGS)
-
-
 ## GCC ###############
 ####### Build Targets #########
 
-searchbench_gcc: GCCCXXFLAGS += -O3 -DNDEBUG
-searchbench_gcc: $(SOURCES) $(HEADERS)
+searchbench: GCCCXXFLAGS += -O3 -DNDEBUG
+searchbench: $(SOURCES) $(HEADERS)
 	$(GCCCXX) $(GCCCXXFLAGS) $(SOURCES) -osearchbench $(LDFLAGS)
 
-debug_gcc : GCCCXXFLAGS += -O0
-debug_gcc : $(SOURCES) $(HEADERS)
+debug : GCCCXXFLAGS += -O0
+debug : $(SOURCES) $(HEADERS)
 		$(GCCCXX) $(GCCCXXFLAGS) $(SOURCES) -odebug $(LDFLAGS)
 
-dump_gcc : src/dump.cc src/benchmark.h
+dump : src/dump.cc src/benchmark.h
 	$(CXX) $(GCCCXXFLAGS) src/dump.cc -odump $(LDFLAGS)
+
+## CLANG ######################
+####### Build Targets #########
+
+clang5_searchbench: CLANGCXXFLAGS += -O3 -DNDEBUG
+clang5_searchbench: $(SOURCES) $(HEADERS)
+	$(CLANGCXX) $(CLANGCXXFLAGS) $(SOURCES) -osearchbench $(LDFLAGS)
+
+clang5_debug: CLANGCXXFLAGS += -O0
+clang5_debug: $(SOURCES) $(HEADERS)
+	$(CLANGCXX) $(CLANGCXXFLAGS) $(SOURCES) -odebug $(LDFLAGS)
+
+clang5_dump: src/dump.cc src/benchmark.h
+	$(CLANGCXX) $(CLANGCXXFLAGS) src/dump.cc -o $dump $(LDFLAGS)
