+++ src/linear_search.h
@@ -7,7 +7,7 @@
 
 template <class Vector, int n = 8> class LinearUnroll {
   template <bool reverse = false>
-  static int64_t linUnroll(const Vector &a, int64_t m, Key k) {
+  static int64_t linearSearchUnroll(const Vector &a, int64_t m, Key k) {
     for (;; m = (reverse ? m - n : m + n)) {
       for (int i = 0; i < n; i++) {
         if (reverse ? (a[m - i] <= k) : (a[m + i] >= k)) {
@@ -19,10 +19,10 @@ template <class Vector, int n = 8> class LinearUnroll {
 
 public:
   static int64_t forward(const Vector &a, const int64_t guessIx, const Key x) {
-    return linUnroll<false>(a, guessIx, x);
+    return linearSearchUnroll<false>(a, guessIx, x);
   }
   static int64_t reverse(const Vector &a, const int64_t guessIx, const Key x) {
-    return linUnroll<true>(a, guessIx, x);
+    return linearSearchUnroll<true>(a, guessIx, x);
   }
 };
