+++ src/sip.h
@@ -1,0 +1,4 @@
+#ifndef SIP_H
+#define SIP_H
+
+#endif //SIP_H
