+++ src/tip.h
@@ -1,0 +1,8 @@
+//
+// Created by Yannis Chronis on 2019-09-12.
+//
+
+#ifndef TIP_H
+#define TIP_H
+
+#endif //TIP_H
